feat: add create butto & relations between User and Todo models

+++ routes/modules/record.js
@@ -14,17 +14,18 @@ router.get('/new', async (req, res) => {
 })
 
 router.post('/', async (req, res) => {
+  const userId = req.user._id
   const { name, date, categoryId, amount } = req.body
   const categories = await Category.find().lean()
   const errors = []
   if (!name || !date || !categoryId || !amount) {
-    errors.push({ message: '所有欄位都是必填' })
+    errors.push({ msg: '所有欄位都是必填' })
   }
   if (checkDate(date) === false) {
-    errors.push({ message: '請輸入合法日期' })
+    errors.push({ msg: '請輸入合法日期' })
   }
   if (isFuture(date) === true) {
-    errors.push({ message: '請勿輸入未來日期' })
+    errors.push({ msg: '請勿輸入未來日期' })
   }
 
   if (errors.length) {
@@ -32,15 +33,16 @@ router.post('/', async (req, res) => {
       errors, name, date, categoryId, amount, categories
     })
   }
-  await Record.create({ name, date, categoryId, amount })
+  await Record.create({ userId, name, date, categoryId, amount })
   res.redirect('/')
 })
 
 //edit 
 router.get('/:id/edit', async (req, res) => {
+  const userId = req.user._id
   const _id = req.params.id
   const categoryData = await Category.find().lean()
-  return Record.findOne({ _id })
+  return Record.findOne({ userId, _id })
     .lean()
     .then((record) => {
       categoryData.filter(category => {
@@ -50,6 +52,7 @@ router.get('/:id/edit', async (req, res) => {
         }
       })
       record.date = moment(record.date).format('YYYY-MM-DD')
+
       return record
     }).then((record) => {
       res.render('edit', { record, categories: categoryData })
@@ -61,18 +64,19 @@ router.get('/:id/edit', async (req, res) => {
 })
 
 router.put('/:id', async (req, res) => {
+  const userId = req.user._id
   const _id = req.params.id
   const { name, date, categoryId, amount } = req.body
   const categories = await Category.find().lean()
   const errors = []
   if (!name || !date || !categoryId || !amount) {
-    errors.push({ message: '所有欄位都是必填' })
+    errors.push({ msg: '所有欄位都是必填' })
   }
   if (checkDate(date) === false) {
-    errors.push({ message: '請輸入合法日期' })
+    errors.push({ msg: '請輸入合法日期' })
   }
   if (isFuture(date) === true) {
-    errors.push({ message: '請勿輸入未來日期' })
+    errors.push({ msg: '請勿輸入未來日期' })
   }
   if (errors.length) {
     return res.render('new', {
@@ -80,7 +84,7 @@ router.put('/:id', async (req, res) => {
     })
   }
 
-  Record.findByIdAndUpdate(_id, req.body)
+  Record.findByIdAndUpdate({ userId, _id }, req.body)
     .then(() => res.redirect('/'))
     .catch(error => {
       console.log(error)
@@ -91,9 +95,10 @@ router.put('/:id', async (req, res) => {
 
 //delate
 router.delete('/:id', (req, res) => {
+  const userId = req.user._id
   const _id = req.params.id
   console.log(_id)
-  Record.findOneAndDelete({ _id })
+  Record.findOneAndDelete({ userId, _id })
     .then(() => res.redirect('/'))
     .catch(error => {
       console.log(error)
